Fixed the double shift bug.
Bug was caused by the IF in line 310. (Line 324 before fixing.)

When a shift was filled, we would reset hasShift of all employees to true. That happened because we coded that condition when we only had 8 hour days.

The program assumed if a shift was full, the day would change. Added a day counter and it's working fine now.