Statistics-sending prototype.
The core of the connection to the database.
Need to create new table in DB for the stats which is connected by a foreign key to the Admin ID (at least I think so), decide what fields it'll contain, match the parameters from Stats.Java line 121 to the new fields, and then code the PHP connection before continuing client-side development.